move SubscribeTask.java to client package
